eliminati bug modifica

+++ ScadenzaDiLegge/Models/DboMarinaresco.cs
@@ -9,20 +9,20 @@ namespace ScadenzaDiLegge.Models
 {
     public partial class DboMarinaresco
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Nave { get; set; }
         public string Comando { get; set; }
         public string Base { get; set; }
-        public long? Visto { get; set; }
+        public int? Visto { get; set; }
         public string TipologiaApparecchiature { get; set; }
         public string ApparecchiaturaSistemazione { get; set; }
         public string Posizione { get; set; }
         public string MarcaModelloDimensioni { get; set; }
         public string TipoDiAccertamento { get; set; }
         public string DataEffettuazione { get; set; }
-        public long? ValiditaAnni { get; set; }
+        public int? ValiditaAnni { get; set; }
         public string ProssimaScadenza { get; set; }
-        public long? GiorniMancantiAllaScadenza { get; set; }
+        public int? GiorniMancantiAllaScadenza { get; set; }
         public string Note { get; set; }
         public string DocumentiCorrelati { get; set; }
         public string Certificati { get; set; }
